Create function to create descriptors from audio recording, satellite type, image..
